src/producer/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
requests==2.31.0